Safari-611.1.10/LayoutTests/http/tests/cache/disk-cache/disk-cache-validation-attachment.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var testMatrix =
[
 [
  { responseHeaders: {'Cache-control': 'max-age=0', 'Content-Disposition': 'attachment; filename="f.txt"' } },
  { responseHeaders: {'Cache-control': 'max-age=100', 'Content-Disposition': 'attachment; filename="f.txt"' } },
  ],
 ];

description("Tests that responses with 'content-disposition: attachment' are cacheable");

var tests = generateTests(testMatrix);

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
