LayoutTests/http/tests/cache/disk-cache/disk-cache-validation-attachment-expected.txt

Tests that responses with 'content-disposition: attachment' are cacheable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 6 tests

response headers: undefined
response source: Network

response headers: {"Cache-control":"max-age=0"}
response source: Network

response headers: {"Cache-control":"max-age=100"}
response source: Disk cache

response headers: {"Content-Disposition":"attachment; filename=\"f.txt\""}
response source: Network

response headers: {"Cache-control":"max-age=0","Content-Disposition":"attachment; filename=\"f.txt\""}
response source: Network

response headers: {"Cache-control":"max-age=100","Content-Disposition":"attachment; filename=\"f.txt\""}
response source: Disk cache

PASS successfullyParsed is true

TEST COMPLETE
